refactor(test): ControllerTest를 E2ETest로 이름 변경
