+++ src/test/java/gift/e2e/PaginationE2ETest.java
@@ -1,4 +1,4 @@
-package gift;
+package gift.e2e;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
